chatbot_rag/requirements.txt
fastapi[standard]
uvicorn
neo4j
python-dotenv
langchain
langchain-openai
pydantic
